4. Тесты для стороны прямоугольника. [#486146]

+++ src/ru/job4j/condition/SqArea.java
@@ -1,12 +1,12 @@
 package ru.job4j.condition;
 
 public class SqArea {
-    public static double square(int p, int k) {
+    public static double square(int p, double k) {
         /*double h = p / (2 * (k + 1));
         double l = h * k;
         double s = l * h;
         return s;*/
-        double h = (double) p / (2 * (k + 1));
+        double h = p / (2 * (k + 1));
         return ((h * k) * h);
     }
